kotlin/foundation: Clarify API around dispatchers
BackgroundExecutingUseCase won't work as advertised unless the scope has
an appropriate dispatcher set up. Avoid confusion by renaming the class
and adding an explicit constructor argument for the dispatcher.
