+++ app-android-journal3/src/main/kotlin/com/hadisatrio/apps/android/journal3/story/EditAStoryActivity.kt
@@ -31,7 +31,7 @@ import com.hadisatrio.libs.android.foundation.widget.EditTextInputEventSource
 import com.hadisatrio.libs.android.foundation.widget.MainThreadEnforcingEventSource
 import com.hadisatrio.libs.android.foundation.widget.TextViewStringPresenter
 import com.hadisatrio.libs.android.foundation.widget.ViewClickEventSource
-import com.hadisatrio.libs.kotlin.foundation.BackgroundExecutingUseCase
+import com.hadisatrio.libs.kotlin.foundation.CoroutineDispatchingUseCase
 import com.hadisatrio.libs.kotlin.foundation.event.CompletionEvent
 import com.hadisatrio.libs.kotlin.foundation.event.EventSources
 import com.hadisatrio.libs.kotlin.foundation.presentation.AdaptingPresenter
@@ -46,8 +46,9 @@ class EditAStoryActivity : AppCompatActivity() {
 
         setContentView(R.layout.activity_edit_a_story)
 
-        BackgroundExecutingUseCase(
-            coroutineScope = lifecycleScope + Dispatchers.Default,
+        CoroutineDispatchingUseCase(
+            coroutineScope = lifecycleScope,
+            coroutineDispatcher = Dispatchers.Default,
             origin = EditAStoryUseCase(
                 targetId = BundledTargetId(intent),
                 stories = journal3Application.stories,
